TS conversion; setting up workspaces to build, test, and demo this project. Goal: build via node v18+, test in Node v10-lts, demo via node v18 in browser.
