+++ .github/workflows/publish.yaml
@@ -16,11 +16,16 @@ jobs:
       - run: |
           git config user.name github-actions
           git config user.email github-actions@github.com
-          npm version patch
+          git fetch
+          git pull origin main
+          npm version patch -ws --include-workspace-root
+          git add .
+          git commit -m "CI - package versions"
+          npm version from-git --allow-same-version
           git push
           git push --tags
-      - run: npm ci
-      - run: npm run build
-      - run: npm publish --access public
+      - run: npm ci -w core
+      - run: npm run build -w core
+      - run: npm publish --access public -w core
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
