+++ .github/workflows/test.yaml
@@ -11,20 +11,21 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node-version: [10.x, 12.x, 14.x, 16.x, 18.x]
+        node-version: [10.x, 12.x, 14.x, 16.x, 18.x, 20.x]
     steps:
     - uses: actions/checkout@v3
+      with:
+        fetch-depth: 0
     - name: Use Node.js 18 (for build)
       uses: actions/setup-node@v3
       with:
         node-version: 18
         cache: 'npm'
-    - run: npm ci
-    - run: npm run build --if-present
+    - run: npm ci -w core
+    - run: npm run build -w core
     - name: Use Node.js ${{ matrix.node-version }} for test runtime
       uses: actions/setup-node@v3
       with:
         node-version: ${{ matrix.node-version }}
         cache: 'npm'
-    - run: npm ci
-    - run: npm test
+    - run: cd test-runner && npm i --omit=dev && npm test
