+++ .gitignore
@@ -20,3 +20,4 @@ Thumbs.db
 .config
 .private
 .npmrc
+*.tsbuildinfo
