--- .mocharc-ts.json
@@ -1,14 +1,0 @@
-{
-  "reporter": "spec",
-  "ui": "bdd",
-  "recursive": true,
-  "colors": true,
-  "exit": true,
-  "require": [
-    "ts-node/register",
-    "globals"
-  ],
-  "extension": [
-    "ts"
-  ]
-}
