+++ .nvmrc
@@ -1,1 +1,1 @@
-hydrogen
+lts/hydrogen
