+++ core/.eslintrc.json
@@ -1,0 +1,64 @@
+{
+	"extends": [
+		"eslint:recommended",
+		"plugin:import/recommended",
+		"plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended"
+	],
+
+	"parser": "@typescript-eslint/parser",
+	
+	"env": {
+		"node": true
+	},
+
+	"plugins": [
+		"import",
+		"node",
+		"@typescript-eslint"
+	],
+
+	"globals": {
+		"document": "readonly",
+		"navigator": "readonly",
+		"window": "readonly",
+		"fetch": "readonly",
+		"google": "readonly",
+		"vi": "readonly"
+	},
+
+	"settings": {
+		"import/resolver": {
+			"node": {
+				"extensions": [".ts",".tsx",".js",".jsx"]
+			}
+		}
+	},
+
+	"rules": {
+		"indent": ["error", "tab", {
+			"SwitchCase": 1,
+			"VariableDeclarator": 1,
+			"outerIIFEBody": 1,
+			"MemberExpression": 1,
+			"FunctionDeclaration": { "parameters": 1, "body": 1 },
+			"FunctionExpression": { "parameters": 1, "body": 1 },
+			"CallExpression": { "arguments": 1 },
+			"ArrayExpression": 1,
+			"ObjectExpression": 1,
+			"ImportDeclaration": 1,
+			"flatTernaryExpressions": false,
+			"ignoreComments": false,
+			"ignoredNodes": ["TemplateLiteral *", "JSXElement", "JSXElement > *", "JSXAttribute", "JSXIdentifier", "JSXNamespacedName", "JSXMemberExpression", "JSXSpreadAttribute", "JSXExpressionContainer", "JSXOpeningElement", "JSXClosingElement", "JSXFragment", "JSXOpeningFragment", "JSXClosingFragment", "JSXText", "JSXEmptyExpression", "JSXSpreadChild"],
+			"offsetTernaryExpressions": true
+		}],
+
+		"@typescript-eslint/no-explicit-any": "off",
+		"@typescript-eslint/no-unused-vars": ["error", {
+			"varsIgnorePattern": "React"
+		}],
+
+		"node/no-missing-import": "off",
+		"node/no-unsupported-features/es-syntax": "off"
+	}
+}
