+++ core/LICENSE.md
@@ -1,4 +1,4 @@
-Copyright (c) 2022, Nathan Bridgewater / Integrated Web Systems, LLC
+Copyright (c) 2023, Nathan Bridgewater / Integrated Web Systems, LLC
 
 All rights reserved.
