+++ core/package.json
@@ -1,0 +1,55 @@
+{
+  "name": "@iwsio/json-csv-core",
+  "version": "0.2.14-alpha.3",
+  "description": "Easily convert JSON to CSV. This is a core library built for browser and node to support buffered conversion to CSV.",
+  "engines": {
+    "node": ">=10"
+  },
+  "main": "dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "./dist"
+  ],
+  "exports": {
+    ".": "./dist/index.js",
+    "./*": "./dist/*.js"
+  },
+  "scripts": {
+    "prebuild": "rimraf dist && tsc --build tsconfig.build.json --clean",
+    "build": "tsc --build tsconfig.build.json"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:IWSLLC/json-csv-core.git"
+  },
+  "keywords": [
+    "csv",
+    "json",
+    "convert",
+    "export csv",
+    "convert csv",
+    "typescript",
+    "export",
+    "buffered",
+    "browser"
+  ],
+  "author": "Nathan Bridgewater <info@iws.io> (https://iws.io/)",
+  "license": "BSD-3-Clause",
+  "bugs": {
+    "url": "https://github.com/IWSLLC/json-csv-core/issues"
+  },
+  "homepage": "https://github.com/IWSLLC/json-csv-core#readme",
+  "devDependencies": {
+    "@types/chai": "^4.3.4",
+    "@typescript-eslint/eslint-plugin": "^5.39.0",
+    "@typescript-eslint/parser": "^5.39.0",
+    "eslint": "^8.41.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-node": "^11.1.0",
+    "npm-run-all": "^4.1.5",
+    "rimraf": "^3.0.2",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3"
+  }
+}
