+++ core/src/index.ts
@@ -1,0 +1,7 @@
+import { buffered } from './exporter'
+export * as exporter from './exporter'
+
+
+export { buffered, buffered as toCsv }
+
+export default buffered
