+++ core/tsconfig.build.json
@@ -1,0 +1,8 @@
+{
+	"extends": "./tsconfig",
+	"compilerOptions": {
+		"sourceMap": false,
+		"declarationMap": false
+	},
+	"exclude": ["node_modules", "**/*.test.*", "**/__tests__/*", "**/__mocks__/*"]
+}
