+++ demo/src/index.html
@@ -1,10 +1,13 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Testing json-csv - TestCafe</title>
-    <script type="text/javascript" src="./dist/index_bundle.js"></script>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta
+      name="description"
+      content="Web site created using create-react-app"
+    />
+    <title>json-csv-core demo</title>
   </head>
   <body>
     <h1>json-csv</h1>
