+++ demo/src/node-advanced.mjs
@@ -1,4 +1,4 @@
-import { toCsv } from '../dist/index.js'
+import { toCsv } from '../../core/dist/index.js'
 import { items, options } from './advanced-setup.mjs'
 
 const csv = toCsv(items, options)
