+++ demo/src/node-simple.mjs
@@ -1,0 +1,44 @@
+import { toCsv } from '../../core/dist/index.js'
+
+const items = [
+	{
+		name: 'fred',
+		email: 'fred@somewhere',
+		amount: 1.02
+	},
+	{
+		name: 'jo',
+		email: 'jo@somewhere',
+		amount: 1.02
+	},
+	{
+		name: 'jo with a comma,',
+		email: 'jo@somewhere',
+		amount: 1.02
+	},
+	{
+		name: 'jo with a quote"',
+		email: 'jo@somewhere',
+		amount: 1.02
+	}
+]
+
+const options = {
+	fields: [
+		{
+			name: 'name',
+			label: 'Name',
+			quoted: true
+		},
+		{
+			name: 'email',
+			label: 'Email'
+		},
+		{
+			name: 'amount',
+			label: 'Amount'
+		}
+	]
+}
+const csv = toCsv(items, options)
+console.log(csv)
