--- src/index.js
@@ -1,5 +1,0 @@
-import { buffered, toCsv, checkOptions, prepValue, getBodyRow, getHeaderRow, getValue, getValueIx } from './exporter'
-
-export { buffered, toCsv, checkOptions, prepValue, getBodyRow, getHeaderRow, getValue, getValueIx }
-
-export default buffered
