+++ test-runner/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@iwsio/json-csv-core-test",
+  "version": "0.2.14-alpha.3",
+  "description": "",
+  "main": "index.js",
+  "engines": {
+    "node": ">=10"
+  },
+  "scripts": {
+    "test": "mocha"
+  },
+  "author": "Nathan Bridgewater",
+  "license": "ISC",
+  "dependencies": {
+    "chai": "^4.3.7",
+    "mocha": "^9.2.2"
+  },
+  "devDependencies": {
+    "eslint": "^8.41.0",
+    "eslint-plugin-chai-friendly": "^0.7.2",
+    "eslint-plugin-mocha": "^10.1.0",
+    "eslint-plugin-node": "^11.1.0"
+  }
+}
