+++ test-runner/test/globals.js
@@ -1,0 +1,3 @@
+const chai = require('chai')
+
+global.expect = chai.expect
