--- test/globals.js
@@ -1,3 +1,0 @@
-import { expect } from 'chai'
-
-global.expect = expect
