+++ tsconfig-base.json
@@ -1,0 +1,22 @@
+
+{
+	"compilerOptions": {
+		"allowJs": true,
+		"allowSyntheticDefaultImports": true,
+		"composite": true,
+		"declaration": true,
+		"declarationMap": true,
+		"esModuleInterop": true,
+		"forceConsistentCasingInFileNames": true,
+		"incremental": true,
+		"inlineSourceMap": false,
+		"lib": ["dom", "dom.iterable", "esnext"],
+		"module": "CommonJS",
+		"moduleResolution": "node16",
+		"resolveJsonModule": true,
+		"skipDefaultLibCheck": true,
+		"skipLibCheck": true,
+		"sourceMap": true,
+		"target": "ES5"
+	}
+}
