--- tsconfig.json
@@ -1,26 +1,0 @@
-{
-  // Change this to match your project
-  "include": ["src/**/*"],
-  "compilerOptions": {
-    // Tells TypeScript to read JS files, as
-    // normally they are ignored as source files
-		"allowJs": true,
-		"target": "ESNext",
-		"module": "NodeNext",
-		"lib": ["ESNext"],
-		"skipLibCheck": true,
-		"moduleResolution": "NodeNext",
-    // Generate d.ts files
-    "declaration": true,
-    // This compiler run should
-    // only output d.ts files
-    "emitDeclarationOnly": true,
-    // Types should go into this directory.
-    // Removing this would place the .d.ts files
-    // next to the .js files
-    "outDir": "dist",
-    // go to js file when using IDE functions like
-    // "Go to Definition" in VSCode
-    "declarationMap": true
-  }
-}
